pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[project]
name = "palate"
version = "0.1.0"
description = "MCP-first personal taste memory and decision engine"
requires-python = ">=3.12"
dependencies = [
  "google-api-python-client>=2.180.0",
  "google-auth-httplib2>=0.2.0",
  "google-auth-oauthlib>=1.2.0",
  "mcp>=1.25.0",
  "openai>=2.14.0",
  "python-dotenv>=1.2.0"
]

[project.scripts]
palate = "palate.server:main"
palate-seed = "palate.seed:main"
palate-google-drive-auth = "palate.google_drive:main"

[tool.setuptools.packages.find]
include = ["palate*"]
